refactor: 컨벤션에 맞게 GlobalStyle 컴포넌트 디렉토리 변경
