+++ src/components/@common/GlobalStyle/GlobalStyle.tsx
